Make image build an explicit step to speed up dev feedback loop

+++ sh/remove_old_images.sh
@@ -13,11 +13,14 @@ remove_all_but_latest()
 {
   local -r docker_image_ls="${1}"
   local -r name="${2}"
+  local -r tag="${CYBER_DOJO_DASHBOARD_TAG}"
   for image_name in `echo "${docker_image_ls}" | grep "${name}:"`
   do
     if [ "${image_name}" != "${name}:latest" ]; then
       if [ "${image_name}" != "${name}:<none>" ]; then
-        docker image rm "${image_name}"
+        if [ "${image_name}" != "${name}:${tag}" ]; then
+          docker image rm "${image_name}"
+        fi
       fi
     fi
   done
